requirements.txt
Django==2.1.1
djangorestframework==3.8.2
gunicorn==19.9.0